added flag for input either Zgas or oh12, not finished
