+++ src/gne_plots.py
@@ -885,6 +885,8 @@ import src.gne_io as io
 #                        plt.close()
 #
 
+
+
 def make_testplots(fnom,zz,verbose=True):    
     # Get output file for BPT plot
     outfile = io.get_outnom(fnom,ftype='plots',ptype='bpt',verbose=verbose)
